stack float instead of int

+++ CPP09/ex01/RPN.hpp
@@ -7,5 +7,5 @@
 
 void    rpn(char *argv);
 void    print_input_error(void);
-void    print_stack(std::stack<int> numbers);
-void    operation(std::stack<int> &numbers, char symbol);
+void    print_stack(std::stack<float> numbers);
+void    operation(std::stack<float> &numbers, char symbol);
